Fixed bug in Buffer function param. Hooked up transaction registry into the ShiDB setup

+++ ShiDB-2.0/src/main/java/buffer/Buffer.java
@@ -56,7 +56,7 @@ public class Buffer {
         contents = new Page(fileMgr.getBlocksize());
     }
 
-    public void setModified(int modifyingTxNum, long lsn) {
+    public void setModified(long modifyingTxNum, long lsn) {
         this.modifyingTxNum = modifyingTxNum;
         if (lsn >= 0)
             this.lsn = lsn;
